+++ tests/projects/zig/console/src/main.zig
@@ -1,0 +1,6 @@
+const std = @import("std");
+
+pub fn main() !void {
+    const stdout = std.io.getStdOut().outStream();
+    try stdout.print("Hello, {}!\n", .{"world"});
+}