+++ tests/projects/zig/console/xmake.lua
@@ -1,0 +1,6 @@
+add_rules("mode.debug", "mode.release")
+
+target("test")
+    set_kind("binary")
+    add_files("src/*.zig")
+