+++ xmake/platforms/linux/xmake.lua
@@ -50,7 +50,7 @@ platform("linux")
     end)
 
     -- set toolchains
-    set_toolchains("envs", "cross", "gcc", "clang", "yasm", "nasm", "fasm", "cuda", "dlang", "go", "rust", "gfortran")
+    set_toolchains("envs", "cross", "gcc", "clang", "yasm", "nasm", "fasm", "cuda", "dlang", "go", "rust", "gfortran", "zig")
 
     -- set menu
     set_menu {