+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### New features
 
 * Add `xmake project -k xcode` generator (use cmake)
-* [870](https://github.com/xmake-io/xmake/issues/870): Support gfortran compiler
+* [#870](https://github.com/xmake-io/xmake/issues/870): Support gfortran compiler
+* [#887](https://github.com/xmake-io/xmake/pull/887): Support zig compiler
 
 ## v2.3.5
 
@@ -779,7 +780,8 @@
 ### 新特性
 
 * 添加xcode工程生成器插件，`xmake project -k cmake` （当前采用cmake生成）
-* [870](https://github.com/xmake-io/xmake/issues/870): 支持gfortran编译器
+* [#870](https://github.com/xmake-io/xmake/issues/870): 支持gfortran编译器
+* [#887](https://github.com/xmake-io/xmake/pull/887): 支持zig编译器
 
 ## v2.3.5